repl: moving fns to internal modules
