+++ lib/internal/readline/utils.js
@@ -439,11 +439,44 @@ function* emitKeys(stream) {
   }
 }
 
+function appendPreview(result, cursorTo, clearScreenDown) {
+  this.previewResult = `\u001b[90m // ${result}\u001b[39m`;
+  const line = `${this._prompt}${this.line} //${result}`;
+  const columns = this.output.columns;
+  const hasColors = this.output.hasColors();
+  const s = hasColors ?
+    `${this._prompt}${this.line}${this.previewResult}` : line;
+
+  // Cursor to left edge.
+  cursorTo(this.output, 0);
+  clearScreenDown(this.output);
+
+  if (columns !== undefined) {
+    this.output.write(line.length < columns ?
+      s : `${s.slice(0, columns - 3)
+        .replace(/\r?\n|\r/g, '')}...\u001b[39m`);
+  } else {
+    this.output.write(s);
+  }
+
+  // Move back the cursor to the original position
+  cursorTo(this.output, this.cursor + this._prompt.length);
+}
+
+function clearPreview() {
+  if (this.previewResult !== '') {
+    this._refreshLine();
+    this.previewResult = '';
+  }
+}
+
 module.exports = {
   emitKeys,
   getStringWidth,
   isFullWidthCodePoint,
   kUTF16SurrogateThreshold,
   stripVTControlCharacters,
-  CSI
+  CSI,
+  appendPreview,
+  clearPreview
 };
