+++ lib/readline.js
@@ -497,37 +497,37 @@ Interface.prototype._insertString = function(c) {
 
 // Append eager eval result to the
 // Current line
-Interface.prototype._appendPreview = function(result) {
-  this.previewResult = `\u001b[90m // ${result}\u001b[39m`;
-  const line = `${this._prompt}${this.line} //${result}`;
-  const columns = this.output.columns;
-  const hasColors = this.output.hasColors();
-  const s = hasColors ?
-    `${this._prompt}${this.line}${this.previewResult}` : line;
-
-  // Cursor to left edge.
-  cursorTo(this.output, 0);
-  clearScreenDown(this.output);
-
-  if (columns !== undefined) {
-    this.output.write(line.length < columns ?
-      s : `${s.slice(0, columns - 3)
-        .replace(/\r?\n|\r/g, '')}...\u001b[39m`);
-  } else {
-    this.output.write(s);
-  }
-
-  // Move back the cursor to the original position
-  cursorTo(this.output, this.cursor + this._prompt.length);
-};
+// Interface.prototype._appendPreview = function(result) {
+//   this.previewResult = `\u001b[90m // ${result}\u001b[39m`;
+//   const line = `${this._prompt}${this.line} //${result}`;
+//   const columns = this.output.columns;
+//   const hasColors = this.output.hasColors();
+//   const s = hasColors ?
+//     `${this._prompt}${this.line}${this.previewResult}` : line;
+//
+//   // Cursor to left edge.
+//   cursorTo(this.output, 0);
+//   clearScreenDown(this.output);
+//
+//   if (columns !== undefined) {
+//     this.output.write(line.length < columns ?
+//       s : `${s.slice(0, columns - 3)
+//         .replace(/\r?\n|\r/g, '')}...\u001b[39m`);
+//   } else {
+//     this.output.write(s);
+//   }
+//
+//   // Move back the cursor to the original position
+//   cursorTo(this.output, this.cursor + this._prompt.length);
+// };
 
 // Refresh the line if preview present
-Interface.prototype._clearPreview = function() {
-  if (this.previewResult !== '') {
-    this._refreshLine();
-    this.previewResult = '';
-  }
-};
+// Interface.prototype._clearPreview = function() {
+//   if (this.previewResult !== '') {
+//     this._refreshLine();
+//     this.previewResult = '';
+//   }
+// };
 
 Interface.prototype._tabComplete = function(lastKeypressWasTab) {
   const self = this;
