+++ lib/repl.js
@@ -74,7 +74,7 @@ const Stream = require('stream');
 const vm = require('vm');
 const path = require('path');
 const fs = require('fs');
-const { Interface } = require('readline');
+const { Interface, cursorTo, clearScreenDown } = require('readline');
 const { Console } = require('console');
 const CJSModule = require('internal/modules/cjs/loader').Module;
 const domain = require('domain');
@@ -90,6 +90,7 @@ const {
   overrideStackTrace,
 } = require('internal/errors');
 const { sendInspectorCommand } = require('internal/util/inspector');
+const { appendPreview, clearPreview } = require('internal/readline/utils');
 const experimentalREPLAwait = require('internal/options').getOptionValue(
   '--experimental-repl-await'
 );
@@ -270,7 +271,7 @@ function REPLServer(prompt,
 
   self[kPreviewResults] = (eagerSession, eagerEvalContextId) => {
     this.on('buffer', (line) => {
-      Interface.prototype._clearPreview.call(this);
+      clearPreview.call(this);
 
       // No need of preview for a multiline statement
       if (this[kBufferedCommandSymbol] !== '')
@@ -291,7 +292,7 @@ function REPLServer(prompt,
 
         if (undefined !== previewResult.result.value) {
           const value = util.inspect(previewResult.result.value);
-          Interface.prototype._appendPreview.call(this, value);
+          appendPreview.call(this, value, cursorTo, clearScreenDown);
           return;
         }
 
@@ -308,8 +309,8 @@ function REPLServer(prompt,
             returnByValue: true,
           }, (err, result) => {
             if (!err) {
-              Interface.prototype._appendPreview
-                .call(this, result.result.value);
+              appendPreview.call(this, result.result.value,
+                                 cursorTo, clearScreenDown);
             }
           });
         }
